Introducing Connector
The Connector interface only expose the remote call stuff.

Also at various places the concrete connector implementations have been replaced by
Connector interface to reduce coupling and simplify testing.
