+++ simulator/src/test/java/com/hazelcast/simulator/protocol/ProtocolIntegrationTest.java
@@ -1,6 +1,7 @@
 package com.hazelcast.simulator.protocol;
 
 import com.hazelcast.simulator.protocol.connector.AgentConnector;
+import com.hazelcast.simulator.protocol.connector.AgentConnectorImpl;
 import com.hazelcast.simulator.protocol.connector.WorkerConnector;
 import com.hazelcast.simulator.protocol.core.Response;
 import com.hazelcast.simulator.protocol.core.ResponseFuture;
